requirements.txt
openai
tenacity
python-dotenv
pytest
pytest-asyncio
ipykernel